requirements.txt
beautifulsoup4
requests
docx2txt
nltk
spacy
jinja2
groq
python-dotenv
pdfplumber